Support larger runs (thread IDs uint16 => uint32)
We hit the limit for thread cardinality in a large object storage
benchmark run (500 hosts, 192 concurrency).  This patch bumps the thread
ID data type from uint16 to uint32 to support larger runs.

Also added an "Assistant workflow for linting" section to CLAUDE.md to
induce efficient auto-formatting and linting.  This, combined with the
added .cursorrules file, helped me with using Cursor with Warp.  The
.cursorrules contents mostly delegate to CLAUDE.md to not repeat tool
guidance/context-compression.

Tweaked the Core Components section of CLAUDE.md to help guide import
understanding.  I had Cursor unnecessarily stumble into circular
dependencies, so this should help with that.

Clarified the CSV format in CLAUDE.md.

+++ cli/benchserver.go
@@ -259,7 +259,7 @@ func runServerBenchmark(ctx *cli.Context, b bench.Benchmark) (bool, error) {
 		case 1:
 			allOps = downloaded[0]
 		default:
-			threads := uint16(0)
+			threads := uint32(0)
 			for _, ops := range downloaded {
 				threads = ops.OffsetThreads(threads)
 				allOps = append(allOps, ops...)
